fixes PLATFORM=newhal build
