+++ hal/src/electron/modem/enums_hal.h
@@ -20,6 +20,7 @@
 #pragma once
 
 #include <string.h>
+#include <stdint.h>
 
 // ----------------------------------------------------------------
 // Types
